// src/constants/enums.ts
export enum LivingCategory {
  HousingCost = "Housing Cost",
  CommunityDemographics = "Community Demographics",
  HealthCare = "Health Care",
  Transportation = "Transportation",
  Education = "Education",
  ParksAndRecreation = "Parks & Recreation",
  GroceriesAndRestaurants = "Groceries & Restaurants",
}

export enum Borough {
  Manhattan = "Manhattan",
  Brooklyn = "Brooklyn",
  Bronx = "Bronx",
  Queens = "Queens",
  StatenIsland = "Staten Island",
}